refactor: IconProps is renamed to IconTypes
