+++ src/assets/Icons/ArrowLeftIcon.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useTheme } from '../../theme/ThemeContext';
 import { colors } from '../../theme/colors';
-import { IconProps } from './IconProps';
+import { IconProps } from './IconTypes';
 
 export const ArrowLeftIcon = ({size}:IconProps) => {
   const theme = useTheme();
