--- src/assets/Icons/IconProps.ts
@@ -1,5 +1,0 @@
-import { iconSizes } from '../../theme/index';
-
-export type IconProps = {
-  size: keyof typeof iconSizes;
-}
