+++ src/assets/Icons/IconTypes.ts
@@ -1,0 +1,26 @@
+import { iconSizes } from '../../theme/index';
+
+export type IconProps = {
+  size: keyof typeof iconSizes;
+}
+
+export type IconName = 
+'ArrowLeftIcon' |
+'CameraIcon' |
+'CheckIcon' |
+'CompassIcon' |
+'DarkIcon' |
+'ErrorIcon' |
+'HeartIcon' |
+'HomeIcon' |
+'LightIcon' |
+'LockIcon' |
+'MobileIcon' |
+'NextIcon' |
+'PauseIcon' |
+'PlayIcon' |
+'PrevIcon' |
+'SearchIcon' |
+'SettingsIcon' |
+'ThemeIcon' |
+'UserIcon';
